Rename sentry config file
